move rewriters to detail
