+++ include/libsemigroups/sims.hpp
@@ -72,11 +72,10 @@
 
 #include <fstream>
 
-#include "debug.hpp"         // for LIBSEMIGROUPS_ASSERT
-#include "exception.hpp"     // for LIBSEMIGROUPS_EXCEPTION
-#include "felsch-graph.hpp"  // for FelschGraph
-#include "knuth-bendix.hpp"  // for KnuthBendix
-#include "libsemigroups/rewriters.hpp"
+#include "debug.hpp"            // for LIBSEMIGROUPS_ASSERT
+#include "exception.hpp"        // for LIBSEMIGROUPS_EXCEPTION
+#include "felsch-graph.hpp"     // for FelschGraph
+#include "knuth-bendix.hpp"     // for KnuthBendix
 #include "presentation.hpp"     // for Presentation, Presentati...
 #include "to-presentation.hpp"  // for to_presentation
 #include "todd-coxeter.hpp"     // for ToddCoxeter
@@ -86,6 +85,7 @@
 #include "matrix.hpp"
 
 #include "detail/iterator.hpp"  // for detail/default_postfix_increment
+#include "detail/rewriters.hpp"
 
 #include "rx/ranges.hpp"
