+++ include/libsemigroups/to-froidure-pin.hpp
@@ -22,13 +22,13 @@
 #include <cstddef>      // for size_t
 #include <type_traits>  // for enable_if_t, is_base_of
 
-#include "debug.hpp"      // for LIBSEMIGROUPS_ASSERT
-#include "exception.hpp"  // for LIBSEMIGROUPS_EXCEPTION
-#include "rewriters.hpp"
+#include "debug.hpp"       // for LIBSEMIGROUPS_ASSERT
+#include "exception.hpp"   // for LIBSEMIGROUPS_EXCEPTION
 #include "word-graph.hpp"  // for WordGraph
 
 #include "detail/kbe.hpp"  // for KBE
 #include "detail/ke.hpp"   // for KE
+#include "detail/rewriters.hpp"
 #include "detail/tce.hpp"  // for TCE
 
 namespace libsemigroups {
