+++ tests/test-knuth-bendix-2.cpp
@@ -73,7 +73,7 @@ namespace libsemigroups {
   using rule_type = KnuthBendix<>::rule_type;
 
 #define KNUTH_BENDIX_TYPES \
-  KnuthBendix<RewriteTrie>, KnuthBendix<RewriteFromLeft>
+  KnuthBendix<detail::RewriteTrie>, KnuthBendix<detail::RewriteFromLeft>
 
   namespace {
     struct weird_cmp {
