+++ tests/test-knuth-bendix-4.cpp
@@ -72,7 +72,7 @@ namespace libsemigroups {
   using rule_type = KnuthBendix<>::rule_type;
 
 #define KNUTH_BENDIX_TYPES \
-  KnuthBendix<RewriteTrie>, KnuthBendix<RewriteFromLeft>
+  KnuthBendix<detail::RewriteTrie>, KnuthBendix<detail::RewriteFromLeft>
 
   ////////////////////////////////////////////////////////////////////////
   // Standard tests
